net/cpfl: fix Coverity issues
Fix integer handling issues, tainted_scalar issues, uninit issues,
overrun issues and control flow issues reported by coverity scan.

Coverity issue: 403259, 403261, 403266, 403267, 403271, 403274
Fixes: db042ef09d26 ("net/cpfl: implement FXP rule creation and destroying")
Fixes: 03f976012304 ("net/cpfl: adapt FXP to flow engine")

Acked-by: Qi Zhang <qi.z.zhang@intel.com>

+++ drivers/net/cpfl/cpfl_ethdev.c
@@ -2478,7 +2478,7 @@ cpfl_alloc_dma_mem_batch(struct idpf_dma_mem *orig_dma, struct idpf_dma_mem *dma
 {
 	int i;
 
-	if (!idpf_alloc_dma_mem(NULL, orig_dma, size * (1 + batch_size))) {
+	if (!idpf_alloc_dma_mem(NULL, orig_dma, (uint64_t)size * (1 + batch_size))) {
 		PMD_INIT_LOG(ERR, "Could not alloc dma memory");
 		return -ENOMEM;
 	}
